Change all Stats to Float, Start Equipment System, make InventoryWindow larger by default
Int-variant of Stat got deleted, stats are floats since using Int doesn't provide known performance benefits.

+++ RoAgain/Assets/Client/Scripts/ClientBattleEntity.cs
@@ -8,7 +8,7 @@ namespace Client
         public int BaseLvl; // for Aura
         public ASkillExecution QueuedSkill;
 
-        public ClientBattleEntity(Coordinate coordinates, LocalizedStringId locNameId, int modelId, float movespeed, int maxHp, int maxSp, int baseLvl, int id = -1) : base(coordinates, locNameId, modelId, movespeed, maxHp, maxSp, id)
+        public ClientBattleEntity(Coordinate coordinates, LocalizedStringId locNameId, int modelId, float movespeed, float maxHp, float maxSp, int baseLvl, int id = -1) : base(coordinates, locNameId, modelId, movespeed, maxHp, maxSp, id)
         {
             BaseLvl = baseLvl;
         }
